requirements.txt
streamlit
transformers
torch
Pillow
sentencepiece
PyMuPDF
regex